refactor(build): fix hard coding of image org (#1703)
There are a few places that had IMAGE_ORG
hardcoded to openebs, making the travis
builds on forked repos to fail.

Also included some minor fixes like:
- added check to copy coverage only if it exists
- removed the step to deploy binaries - which are not used

(cherry picked from commit 0b208c2c0b010581bbc2f0d03f6ec0012dfd227d)

+++ buildscripts/travis-build.sh
@@ -72,7 +72,7 @@ elif [ "$TRAVIS_CPU_ARCH" == "arm64" ]; then
   rc=$?; if [[ $rc != 0 ]]; then exit $rc; fi
 fi
 
-if [ $SRC_REPO != $DST_REPO ];
+if [ $SRC_REPO != $DST_REPO ] && [ -f "coverage.txt" ];
 then
 	echo "Copying coverage.txt to $SRC_REPO"
 	cp coverage.txt $SRC_REPO/
